API-1012: Extract code into a migration class
