+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/install.yml
@@ -12,3 +12,9 @@ services:
             - '@database_connection'
         tags:
             - { name: kernel.event_subscriber }
+
+    Akeneo\Connectivity\Connection\Infrastructure\Install\MigrateAudit40Master:
+        class: Akeneo\Connectivity\Connection\Infrastructure\Install\MigrateAudit40Master
+        arguments:
+            - '@database_connection'
+            - '@akeneo_connectivity.connection.application.audit.handler.update_product_event_count'
